Changed ViewModels code and improved logic.

+++ app/src/main/java/com/example/android/leaguestats/interfaces/IdClickListener.java
@@ -1,5 +1,5 @@
 package com.example.android.leaguestats.interfaces;
 
 public interface IdClickListener {
-    void onClickListener(long id);
+    void onClickListener(int id);
 }
